requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
pandas>=2.2
python-dotenv
streamlit>=1.38
pyarrow
matplotlib>=3.8
seaborn>=0.13